+++ spec/integration/i_spec.rb
@@ -5,9 +5,4 @@ require 'railway_ipc'
 
 Dir[File.dirname(__FILE__) + '/support/**/*.rb'].sort.each { |f| require f }
 
-logger = Logger.new(STDOUT)
-logger.level = :fatal
-
-RailwayIpc.configure(
-  logger: logger
-)
+RailwayIpc.configure(IO::NULL)