+++ spec/spec_helper.rb
@@ -25,7 +25,7 @@ end
 
 RSpec::Matchers.define_negated_matcher :not_change, :change
 
-RailwayIpc.configure(logger: Logger.new(IO::NULL))
+RailwayIpc.configure(IO::NULL)
 Sneakers.logger = Logger.new(IO::NULL)
 
 RSpec.configure do |config|